Rewrite README to give steps for getting project running; fix arrow svg bug since it bothered me
