+++ README.md
@@ -1,36 +1,43 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+<p align="center">
+  <img src="./public/bp_logo_white.svg" alt="Blueprint Logo" width="200"/>
+</p>
 
-## Getting Started
+# UofT Blueprint Website v2
 
-First, run the development server:
+## Installation and Set Up
 
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
+### First clone the repository and move into the directory
+
+```
+git clone https://github.com/uoftblueprint/website-v2.git
+
+cd website-v2
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+### Install Node.js
 
-You can start editing the page by modifying `app/page.js`. The page auto-updates as you edit the file.
+Make sure you have `node.js` installed with at least `v18.20`. If not, install `node.js` (Recommended latest `v22.9`)
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
+**Install for any operating systems**
 
-## Learn More
+Follow the prompts for your system
 
-To learn more about Next.js, take a look at the following resources:
+https://nodejs.org/en/download/package-manager
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+### Install packages
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+After you have `node` installed, run the following in the project root directory
 
-## Deploy on Vercel
+```
+npm install
+```
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+All packages should now be installed
 
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+### Run the Project
+
+The following will create a development instance of the website
+
+```
+npm run dev
+```
