+++ src/app/home-page-top-section/BottomSection.css
@@ -59,7 +59,7 @@
 }
 
 .arrow {
-  width: 70%;
+  width: 25px;
 }
 
 @media (max-width: 992px) {
