+++ src/app/layout.tsx
@@ -1,13 +1,12 @@
 "use client";
 import { useState } from "react";
-import { Poppins } from "next/font/google";
+import { Inter } from "next/font/google";
 import "@mantine/core/styles.layer.css";
 import { MantineProvider, AppShell } from "@mantine/core";
 import Navbar from "@/components/Navbar/Navbar";
 import SideNavMenu from "@/components/Navbar/SideNavMenu";
-import Footer from "@/components/Footer/Footer"
 
-const poppins = Poppins({ weight: "400", subsets: ["latin"] });
+const inter = Inter({ subsets: ["latin"] });
 
 const links = [
   { link: "/about", label: "About" },
@@ -21,15 +20,15 @@ export default function RootLayout({ children }) {
 
   return (
     <html lang="en">
-      <body className={poppins.className}>
+      <body className={inter.className}>
         <MantineProvider>
           <AppShell
             header={{ height: "96px" }}
-            footer={{ height: "10vw" }} // adjust based on footer height
+            footer={{ height: "283px" }} // adjust based on footer height
             style={{ position: "relative" }}
             navbar={{
               width: 300,
-              breakpoint: "md",
+              breakpoint: "sm",
               collapsed: { mobile: !isSideNavOpened, desktop: true },
             }}
             transitionDuration={0}
@@ -49,7 +48,7 @@ export default function RootLayout({ children }) {
             <AppShell.Main>{children}</AppShell.Main>
 
             <AppShell.Footer style={{ position: "absolute" }}>
-              {<Footer></Footer>}
+              {/* add footer here! */}
             </AppShell.Footer>
           </AppShell>
         </MantineProvider>
