--- src/app/page.js
@@ -1,7 +1,0 @@
-"use client";
-
-export default function Home() {
-    return (
-      <></>
-    );
-}
